+++ src/main/java/blackjack/dto/PlayerProfitResult.java
@@ -1,4 +1,4 @@
 package blackjack.dto;
 
-public record PlayerProfitResult(String name, double profit) {
+public record PlayerProfitResult(String name, int profit) {
 }
